rework now-playing

+++ modules/socradio/events.js
@@ -1,7 +1,8 @@
 
 let running = false
-let radioChannel
-let message
+let radioChannel, message
+const axios = require('axios')
+const stations = {}
 
 module.exports = {
   events: {
@@ -20,55 +21,42 @@ module.exports = {
       await Promise.all(messages.map(m => m.delete()))
 
       var socket = require('socket.io-client')('https://api.squid-radio.net')
-      socket.on('metadata', async (data) => {
-        console.log([
-          {
-            name: 'Album',
-            value: data.album,
-            inline: true
-          },
-          {
-            name: 'Artist',
-            value: data.artist,
-            inline: true
-          },
-          {
-            name: 'Track',
-            value: data.title,
-            inline: true
-          }
-        ]
-        )
-        const newMessage = await channel.send({
-          embed: {
-            color: 1719241,
-            thumbnail: {
-              url: encodeURI(`https://squid-radio.net/covers/${data.album}.jpg`)
-            },
-            title: 'Now Playing',
-            url: 'https://play.squid-radio.net/clouds',
-            fields: [
-              {
-                name: 'Album',
-                value: data.album,
-                inline: true
-              },
-              {
-                name: 'Artist',
-                value: data.artist,
-                inline: true
-              },
-              {
-                name: 'Track',
-                value: data.title,
-                inline: true
+
+      axios.get('https://api.squid-radio.net/stations')
+        .then(res => {
+          res.data.forEach(station => {
+            stations[station] = { title: '', artist: '', album: '' }
+            socket.on(station, async (data) => {
+              if (data !== null) {
+                stations[station] = data
+
+                const newMessage = await channel.send({
+                  embed: {
+                    color: 1719241,
+                    thumbnail: {
+                      url: `https://squid-radio.net/images/station/station_${station}.png`
+                    },
+                    title: 'Now Playing',
+                    url: 'https://squid-radio.net',
+                    fields: Object.keys(stations).map(stationName => {
+                      return {
+                        name: capitalize(stationName),
+                        value: `${stations[stationName].album} - ${stations[stationName].artist} - ${stations[stationName].title}`,
+                        inline: true
+                      }
+                    })
+                  }
+                })
+                if (message) message.delete()
+                message = newMessage
               }
-            ]
-          }
+            })
+          })
         })
-        if (message) message.delete()
-        message = newMessage
-      })
+        .catch(err => {
+          console.log('failed to fetch stations')
+          console.log(err)
+        })
     },
     async voiceStateUpdate (client, db) {
       const members = radioChannel.members.filter(m => m.id !== m.guild.me.id)
@@ -86,4 +74,9 @@ module.exports = {
       }
     }
   }
+}
+
+const capitalize = (s) => {
+  if (typeof s !== 'string') return ''
+  return s.charAt(0).toUpperCase() + s.slice(1)
 }
